Added the option of transfer functions rotation to the orthogonal geographic coordinate system to z2xml (based on EMTF/doc/PDF/Z_files.pdf document, and functionally identical to EMTF/matlab/ZPLT/corChangeZ.m).
The declination should not be output back to Z-files: it is only there to
make sure that channel orientations are correct, i.e. true channel orientations
are orientation + declination. In our programs, we set the orientations to the
correct values when we read them first, so declination should be zero in Z-files.

Also, changed the schema to allow for additional information such as the original
MeasurementCoordinates - necessary if we want these transformations to be reversible.
Also, elements can now be written in any order. In general, this schema is much more
flexible. It is not in SPADE yet: might need further changes.

Now, need to update FoX (and subroutines throughout our code); then need to make
sure that information such as MeasurementCoordinates is written out to XML files
and that it is not read in instead of the correct channel orientations. Then, need
to also update z2edi and xml2edi (?), could also write z2z just for the rotations.