FoX updated finally. Programs compile but do not run with the new FoX libraries. Debugging required. Last working version is revision 20.
Still need to make sure that information such as MeasurementCoordinates is written out 
to XML files and that it is not read in instead of the correct channel orientations. 
Could also update z2edi and xml2edi (?); could also write z2z just for the rotations.

Most importantly, interface with FoX has to be updated to restore the functionalities
of the codes.