FCU 2.0
Lots of information added to the lists, including
1) data quality information (site list),
2) field comments (run list), and
3) instruments used to collect the data (channels list).

These fields are now read and used to produce the final TF in SPUD XML format.
The old SPADE XML format is now deprecated.

No schema exists for the SPUD XML (yet), but it's a cleaner, more accurate
and logical format, that's also easier to read.

FCU tools that read the new XML have not yet been updated/developed.

So far, tested on USArray 2011 data. Will now reproduce the other lists.