Starting to work on updating the FCU code to convert all of the existing MT transfer functions to XML and upload them to SPUD. The first issue to sort out is the copyright. Starting some correspondence about this with IRIS (see below).
In the meanwhile, making some obvious changes:

1) Added an optional 'datum' attribute to the Location. Had a discussion
with Randy Mackie in Sept 2012, and this is what he said.

In regards to datums, I am no expert, so I asked someone who does this stuff all the time, and this is what he said:

=====================================>

It depends on what datum. In CA between NAD27 and WGS84, the difference in
geographic coordinates is about 4" in longitude and much lower in latitude
or about 100 m difference, mostly in longitude. 

There are many worse cases. The worst that I've run into with geographic
coordinates is about a 700 m difference, usually in longitude (due to the
greater difficulty of fixing longitude in a datum defined using ship's
clocks and sun elevations). 

There is a widespread conviction that using geographic coordinates absolves
practitioners from specifying datum, leading to lots of errors.

The datum difference in CA between NAD27 and WGS84 in UTM is about 200 m in
Northing and 20 to 80 m in Easting, so projection also matters.

<=====================================

So, probably doesn't matter much for mantle studies, but just as good practice, I think it's best
to always specify the datum.

This is from Wikipedia -

WGS84 - common standard global datum
NAD83 - used in North America
ETRS89 - used in Europe

For example, in Sydney there is a 200 m  difference between GPS coordinates configured in GDA (based on global standard WGS84) and AGD (used for most local maps)

By default using 'WGS84' for the data collected with GPS information in the US.

2) Added a Quality Warning comment and a binary flag. If the flag==1 a warning is provided
in addition to the data quality assessment. This was necessitated by this years USArray data
in the Mid-Continent Ridge: many of the transfer are smooth and of good quality, but they have
a bump at around 1 min period that is clearly a large-scale source of cultural effect which
should not be directly interpreted through an inversion. The warning is therefore designed
to act as a zero-order solution and provide some guidance to people with regards to sites that
may need additional attention of correction before they may be directly interpreted.

3) Starting on a Copyright element. Once ready, it should have the following structure

  <Copyright>
    <Citation>
      <Author>Adam Schultz</Author>
      <Author>Gary D. Egbert</Author>
      <Author>Anna Kelbert</Author>
      <Year>2012</Year>
      <Name>USArray TA Magnetotelluric Transfer Functions</Name>
      <DOI>Unassigned</DOI>
    </Citation>
    <ReleaseStatus>Unrestricted Release</ReleaseStatus>
    <ConditionsOfUse>All data and metadata for this survey are available free of charge 
and may be copied freely, duplicated and further distributed provided this data set 
is cited as the reference. 
Guidelines about the quality or limitations of the data and metadata as obtained from
the authors are included for informational purposes only.
Please refer to the IRIS data use policy for further details.</ConditionsOfUse>
  </Copyright>

Quoting my emails to IRIS as of today:

By default, the citation name would be composed of the Project, Survey and Product Description (as above), unless the author prefers to specify a journal publication.

A full XML example is attached. Please give feedback.

Of course it would be much preferable that any data identifiers (DOI or suchlike) are obtained before the data are published, so that they could be included in the XML files. Otherwise, it would take some effort to regenerate all files once the DOI is obtained. Certainly such practice would also make many people much more willing to share data.

Any chance something like that could be done - DOI's obtained for each of my surveys - in the time frame of a few months? (see NSF Dear Colleague Letter as of Marth 29, 2012, attached). 

PS Data DOI's are encouraged by the NSF, are already used internationally e.g. by NERC (through DataCite.org) and are of course overall a great practice for authorship credit and reverse feedback mechanisms.

Use case: Dr. Wicked wants to publish an analysis of my data before I do. They still have to refer to the data DOI so I get some credit for collecting the data.

Use case: Dr. Clever finds an error in a dataset. He is able to track everyone who's used that dataset through the DOI and let them know. The authors then publish errata to their papers.

Use case: Dr. Young is trying to determine a particular piece of metadata that I thought was too obvious to record when the data were collected. Through the DOI, they find me in the system and get my updated email address. 

---
Will wait and see where this goes.

Next step is to implement the reading of the EDI files.