Finally, was able to clean up and get rid of the explicit dimensions structure. All dimensions (number of frequencies, input and output electric and magnetic channels, and the number of data types) are now treated cleanly through subroutine interfaces, as well as stored as necessary in the corresponding read/write modules.
Also, completed the general Z-file reading and writing. All looks reasonable and seems
to work fine.

All tools are now functioning: z2xml, xml2z, z2edi, edi2xml, xml2edi, and z2z
(the latter may be used to rotate the transfer functions).

The only components of the system that are not completely general at present and won't
work with arbitrary number of channels and data types, are
1) write_edi_file
2) rotate_z_period

EDI writing routines could (and should!) quite easily be made completely generic, working
with any arbitrary data types. Everything is in place to do this, just haven't implemented yet.
Currently, EDI write uses a rewrite of an old subroutine by Randie Mackie and has a bunch of
unnecessary restrictions. The channels are not written quite correctly, either.

Obviously, more work will be needed to accommodate the whole range of EDI files,
but this framework should allow for this.