FCU 3.0 bug fix
1. Fixed an important bug in XML list reading (the matching runs were chosen based on the last 
few characters in the ID being the same; that didn't work for some surveys). 

Known problem: if &apos; occurs at the start of an XML element content - which sometimes happens 
when Matlab writes the list files - FoX library crashes the program.
In this case, edit the list manually to remove the offending character.

2. Updated DATATYPES wherever they were not usable yet; added SeeAlso elements to get rid of
the annoying notifications when these XML files are read.