Some fixes and clarifications to z_read.f90. Added the option to save the header lines for future output; clarified that the "declination" in Z-files is not in fact the declination but orientation relative to geographic North, so that all channel orientations are specified relative to that value.
The major change relates to the computation of the TF variance from the two covariance matrices in Z-files.
I realized that the lines

    TFVar=0.0d0
    do i=1,nchout
       do j=1,nchin
          TFVar(i,j) = (ResidCov(i,i)*InvSigCov(j,j))/2
       end do
    end do

in fact computed the variance of a real or imaginary part of the impedance, not the total complex variance.
However, p 52 of the EDI manual explains that e.g. ZXY.VAR refers to the complex variance 
(while ZXYR.VAR would mean the variance of the real part).

Evidently, I missed this when I was writing the code. To be consistent with this definition, 
we are no longer dividing by two:

    TFVar=0.0d0
    do i=1,nchout
       do j=1,nchin
          TFVar(i,j) = (ResidCov(i,i)*InvSigCov(j,j))
       end do
    end do

This will change the meaning of the variance found in XML files. All XML files currently in the database
will be regenerated using the updated code.