Major organizational fixes an improvements that enable well-defined data citation information to be stored at the SPUD database.
1. Introduction of SurveyDOI element.

I create this MANUALLY in the config.xml file. The complete Citation now looks like:

   <Citation>
      <Title>USArray TA Magnetotelluric Transfer Functions</Title>
      <Authors>Schultz, A., G. D. Egbert, A. Kelbert, T. Peery, V. Clote, B. Fry, S. Erofeeva and staff of the National Geoelectromagnetic Facility and their contractors</Authors>
      <Year>2006-2018</Year>
      <SurveyDOI>doi:10.17611/DP/EMTF/USARRAY/TA</SurveyDOI>
   </Citation>

By agreement with IRIS, SurveyDOI always starts with doi:10.17611/DP/EMTF/ and is completed using unambiguous Project and Survey abbreviations.

In their system, SurveyDOI gets semi-automatically created the first time that a file with this SurveyDOI is submitted to IRIS SPUD. It is critical that no mistakes are made:
once a DOI is created, it is there for good. 

The site DOIs are automatically minted upon submission of each file.

This Citation and some additional info get parsed and combined for presentation as follows:

Survey Reference:	

Schultz, A., G. D. Egbert, A. Kelbert, T. Peery, V. Clote, B. Fry, S. Erofeeva and staff of the National Geoelectromagnetic Facility and their contractors (2006-2018). 
"USArray TA Magnetotelluric Transfer Functions". doi:10.17611/DP/EMTF/USARRAY/TA (Retrieved from the IRIS database on Feb 26, 2016)

Specific Site Reference:	

Schultz, A., G. D. Egbert, A. Kelbert, T. Peery, V. Clote, B. Fry, S. Erofeeva and staff of the National Geoelectromagnetic Facility and their contractors (2006-2018). 
"USArray TA Magnetotelluric Transfer Functions: ORF10". doi:10.17611/DP/2460978 (Retrieved from the IRIS database on Feb 26, 2016)

2. Introduction of Acknowledgement and SelectedPublications elements.

The data use requirements have now been modified to, generally, require that the Acknowledgement element is included in publications that make use of the data.

There is, in general, no requirement that the Selected Publications are cited: these are intended for user convenience only. However, some data authors (read Alan Jones)
do require that their papers are cited. For these occasions, a new release status has been created, called Paper Citation Required. If this condition applies,
it will be stated verbatim in the Conditions of Use.

Three text files may be supplied to provide this information, as well as additional information and acknowledgements of informational character.

    <THANKS>../USArrayTA_acknowledgements.txt</THANKS>
    <PAPERS>../USArrayTA_selectedpubs.txt</PAPERS>
    <README>../USArrayTA_readme.txt</README>

These correspond to the file names in the config.xml file. Their contents populate Acknowledgement, SelectedPublications, and AdditionalInfo elements, respectively.

There is only one caveat so far: if the first line in a text file is empty, this file will be considered empty by the code, and its contents will not be included.
Perhaps that's a "feature" that should be fixed, but it doesn't seem to be a high priority issue.